-Refactor UI code and apply PEP8 style
