+++ core/tester.py
@@ -1,4 +1,5 @@
 import os
+
 import glfw
 import numpy as np
 from core.policy import build_policy
