+++ envs/flamingo_p_v0/flamingo_p_v0.py
@@ -3,12 +3,13 @@ from gymnasium.envs.mujoco import MujocoEnv
 from gymnasium.spaces import Box
 import numpy as np
 import mujoco
+import glfw
 from envs.flamingo_p_v0.manager.control_manager import ControlManager
 from envs.flamingo_p_v0.manager.xml_manager import XMLManager
 from envs.flamingo_p_v0.utils.math_utils import MathUtils
 from envs.flamingo_p_v0.utils.mujoco_utils import MuJoCoUtils
 from envs.flamingo_p_v0.utils.noise_generator_utils import truncated_gaussian_noisy_data, uniform_noisy_data
-import glfw
+
 
 
 class FlamingoPV0(MujocoEnv, utils.EzPickle):
