+++ envs/wheeldog_p_v0/wheeldog_p_v0.py
@@ -3,12 +3,12 @@ from gymnasium.envs.mujoco import MujocoEnv
 from gymnasium.spaces import Box
 import numpy as np
 import mujoco
+import glfw
 from envs.wheeldog_p_v0.manager.control_manager import ControlManager
 from envs.wheeldog_p_v0.manager.xml_manager import XMLManager
 from envs.wheeldog_p_v0.utils.math_utils import MathUtils
 from envs.wheeldog_p_v0.utils.mujoco_utils import MuJoCoUtils
 from envs.wheeldog_p_v0.utils.noise_generator_utils import truncated_gaussian_noisy_data, uniform_noisy_data
-import glfw
 
 
 class WheelDogPV0(MujocoEnv, utils.EzPickle):
