+++ ui/main_window.py
@@ -1,595 +1,20 @@
 import os
 import yaml
+
 from PyQt5.QtWidgets import (
     QWidget, QVBoxLayout, QHBoxLayout, QGridLayout, QFormLayout,
     QPushButton, QLabel, QMessageBox, QMainWindow,
-    QFileDialog, QGroupBox, QScrollArea, QComboBox, QLineEdit, QSlider, QCheckBox, QDialog, QDialogButtonBox, QSizePolicy, QApplication, QStyle
+    QFileDialog, QGroupBox, QScrollArea, QLineEdit, QCheckBox, QDialog
 )
-from PyQt5.QtCore import QThread, pyqtSignal, QObject, Qt, QEvent, QUrl, QTimer
-from PyQt5.QtGui import QDesktopServices, QIcon, QDoubleValidator, QIntValidator
+from PyQt5.QtCore import QThread, Qt, QEvent, QUrl
+from PyQt5.QtGui import QDesktopServices, QIcon
 from core.tester import Tester
+from ui.utils import to_float, to_int, normalize_numkey_float_values
+from ui.custom_widgets import NoWheelComboBox, NoWheelSlider, NonClickableButton
+from ui.dialogs.hardware_settings import HardwareSettingsDialog
+from ui.dialogs.observation_settings import ObservationSettingsDialog
+from ui.workers import TesterWorker
 
-# ---------------------------
-# Utility Widgets
-# ---------------------------
-
-class NoWheelComboBox(QComboBox):
-    # Ignore mouse wheel to prevent accidental selection changes (event ignored -> bubble up to parent, enabling scroll area to catch it)
-    def wheelEvent(self, event):
-        event.ignore()
-
-class NoWheelSlider(QSlider):
-    # Ignore mouse wheel to prevent accidental value changes
-    def wheelEvent(self, event):
-        event.ignore()
-
-class NonClickableButton(QPushButton):
-    # Make the button non-interactive to mouse clicks (visual indicator only)
-    def mousePressEvent(self, event):
-        event.ignore()
-    def mouseReleaseEvent(self, event):
-        event.ignore()
-
-# ---------------------------
-# Helpers: safe casting & normalization
-# ---------------------------
-
-def to_float(val, default=1.0):
-    # Safe float conversion with fallback
-    try:
-        return float(val)
-    except Exception:
-        return float(default)
-
-def to_int(val, default=0):
-    # Safe int conversion with fallback
-    try:
-        return int(val)
-    except Exception:
-        return int(default)
-
-def normalize_numkey_float_values(d):
-    """
-    For dictionaries like command_scales, even if keys come as 0 or '0',
-    normalize keys to strings ('0'..'5') and ensure values are floats.
-    """
-    if not isinstance(d, dict):
-        return {}
-    out = {}
-    for k, v in d.items():
-        out[str(k)] = to_float(v, 1.0)
-    return out
-
-# ---------------------------
-# Hardware Settings Dialog
-# ---------------------------
-
-class HardwareSettingsDialog(QDialog):
-    def __init__(self, hardware_settings, parent=None):
-        super().__init__(parent)
-        self.hardware_settings = (hardware_settings or {}).copy()
-        self.setWindowTitle("Hardware Settings")
-        self._setup_ui()
-
-    def _setup_ui(self):
-        # Main vertical layout for the dialog
-        main_layout = QVBoxLayout(self)
-
-        # Create a scroll area to allow scrolling when height exceeds limit
-        scroll = QScrollArea()
-        scroll.setWidgetResizable(True)  # Allow resizing of the scroll area content
-        scroll.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)  # Hide horizontal scrollbar
-
-        # Create an inner widget that will hold the form layout
-        inner_widget = QWidget()
-        form_layout = QFormLayout(inner_widget)
-        self.fields = {}
-
-        # Build editable fields for each hardware setting
-        for key, value in self.hardware_settings.items():
-            label = QLabel(key)
-            le = QLineEdit(str(value))
-            le.setValidator(QDoubleValidator())  # Allow only floating-point input
-            form_layout.addRow(label, le)
-            self.fields[key] = le
-
-        # Set the inner widget inside the scroll area
-        scroll.setWidget(inner_widget)
-        main_layout.addWidget(scroll)
-
-        # OK / Cancel buttons at the bottom
-        buttons = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        buttons.accepted.connect(self.accept)   # Close dialog with "Accepted" state
-        buttons.rejected.connect(self.reject)   # Close dialog with "Rejected" state
-        main_layout.addWidget(buttons)
-
-        # Set maximum height to 600px; scrolling will be enabled if exceeded
-        self.setMaximumHeight(600)
-
-    def get_settings(self):
-        # Return current field values as a dictionary
-        return {key: le.text() for key, le in self.fields.items()}
-
-# ---------------------------
-# Observation Settings Dialog (SCROLLABLE)
-# ---------------------------
-
-class ObservationSettingsDialog(QDialog):
-    def __init__(self, settings, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Observation Settings")
-        self.obs_types = ["dof_pos", "dof_vel", "ang_vel", "lin_vel_x", "lin_vel_y", "lin_vel_z", "projected_gravity", "height_map", "last_action"]
-        self.settings = settings if isinstance(settings, dict) else {}
-        self.parent_widget = parent
-
-        # Row stores for stacked / non-stacked entries
-        self.stacked_rows = []   # dicts: {"layout": QHBoxLayout, "combo": QComboBox, "freq": QComboBox, "scale": QComboBox}
-        self.non_rows = []
-
-        # Command scale widgets
-        self.cmd_scale_cbs = []
-        self.command_scales_grid = None
-
-        self._setup_ui()
-
-    def _scale_options(self):
-        # Predefined numeric options for scale combo boxes
-        return ["0", "0.01", "0.05", "0.1", "0.15", "0.25", "0.5", "0.75", "1.0", "2.0", "2.5", "5"]
-
-    def _setup_ui(self):
-        # ====== Outer layout for the dialog ======
-        self._outer_layout = QVBoxLayout(self)
-
-        # ====== Scroll area wrapping the inner content ======
-        self._scroll = QScrollArea()
-        self._scroll.setWidgetResizable(True)  # Make the scroll area adapt to dialog resizing
-        self._scroll.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)  # Avoid horizontal scrolling
-        self._outer_layout.addWidget(self._scroll)
-
-        # ====== Inner widget & its main layout (everything scrolls inside here) ======
-        self._inner_widget = QWidget()
-        self._inner_layout = QVBoxLayout(self._inner_widget)
-
-        # ---- obs label ----
-        obs_label = QLabel("State Order: [Stacked]  →  [Non-Stacked]  →  [Command]")
-        obs_label.setAlignment(Qt.AlignCenter)
-        obs_label.setStyleSheet("""
-            QLabel {
-                font-size: 7pt;
-                color: #222222;
-                letter-spacing: 1px;
-            }
-        """)
-        obs_label.setSizePolicy(QSizePolicy.Preferred, QSizePolicy.Fixed)
-
-        self._inner_layout.addWidget(obs_label)
-
-        # ---- Load env config & current settings (used throughout UI building) ----
-        env_id = self.parent_widget.env_id_cb.currentText()
-        env_cfg = self.parent_widget.env_config.get(env_id, {}) or {}
-        cmd_cfg_raw = env_cfg.get("command", {}) if isinstance(env_cfg.get("command", {}), dict) else {}
-        obs_scales = env_cfg.get("obs_scales", {}) if isinstance(env_cfg.get("obs_scales", {}), dict) else {}
-        command_scales_from_cfg = normalize_numkey_float_values(env_cfg.get("command_scales", {}))
-        default_stack_size = to_int(env_cfg.get("stack_size", self.settings.get("stack_size", 3)), 3)
-
-        # ----- Prefer existing settings over YAML defaults -----
-        cmd_dim_val = to_int(self.settings.get("command_dim", cmd_cfg_raw.get("command_dim", 6)), 6)
-
-        # ---------------- Stacked Observation ----------------
-        stacked_group = QGroupBox("Stacked Observation")
-        stacked_v = QVBoxLayout()
-
-        stack_size_h = QHBoxLayout()
-        stack_size_h.addWidget(QLabel("Stack Size:"))
-        self.stack_size_cb = NoWheelComboBox()
-        self.stack_size_cb.addItems([str(i) for i in range(1, 16)])
-        self.stack_size_cb.setCurrentText(str(self.settings.get("stack_size", default_stack_size)))
-        stack_size_h.addWidget(self.stack_size_cb)
-        stacked_v.addLayout(stack_size_h)
-
-        self.stacked_container = QVBoxLayout()
-        stacked_v.addLayout(self.stacked_container)
-        add_btn = QPushButton("Add")
-        add_btn.clicked.connect(lambda: self.add_stacked())
-        stacked_v.addWidget(add_btn)
-        stacked_group.setLayout(stacked_v)
-
-        # ---------------- Non-Stacked Observation ----------------
-        non_group = QGroupBox("Non-Stacked Observation")
-        non_v = QVBoxLayout()
-        self.non_container = QVBoxLayout()
-        non_v.addLayout(self.non_container)
-        add_non = QPushButton("Add")
-        add_non.clicked.connect(lambda: self.add_non())
-        non_v.addWidget(add_non)
-        non_group.setLayout(non_v)
-
-        # Two groups side-by-side
-        areas_layout = QHBoxLayout()
-        areas_layout.addWidget(stacked_group)
-        areas_layout.addWidget(non_group)
-        self._inner_layout.addLayout(areas_layout)
-
-        # ---------------- Height Map (size/res fields) ----------------
-        height_group = QGroupBox("Height Map")
-        height_layout = QFormLayout()
-        size_res_layout = QHBoxLayout()
-
-        hm_yaml = env_cfg.get("height_map", {}) if isinstance(env_cfg.get("height_map", {}), dict) else {}
-        hm_settings = self.settings.get("height_map", {}) if isinstance(self.settings.get("height_map", {}), dict) else {}
-        hm_default = {
-            "size_x": to_float(hm_settings.get("size_x", hm_yaml.get("size_x", 1.0))),
-            "size_y": to_float(hm_settings.get("size_y", hm_yaml.get("size_y", 0.6))),
-            "res_x": to_int(hm_settings.get("res_x", hm_yaml.get("res_x", 15))),
-            "res_y": to_int(hm_settings.get("res_y", hm_yaml.get("res_y", 9))),
-        }
-
-        self.height_size_x_le = QLineEdit(str(hm_default["size_x"]))
-        self.height_size_x_le.setFixedWidth(50)
-        self.height_size_x_le.setPlaceholderText("x (m)")
-        double_validator = QDoubleValidator(0.000001, 1e6, 4)
-        double_validator.setNotation(QDoubleValidator.StandardNotation)
-        self.height_size_x_le.setValidator(double_validator)
-        size_res_layout.addWidget(self.height_size_x_le)
-        size_res_layout.addWidget(QLabel("×"))
-        self.height_size_y_le = QLineEdit(str(hm_default["size_y"]))
-        self.height_size_y_le.setFixedWidth(50)
-        self.height_size_y_le.setPlaceholderText("y (m)")
-        self.height_size_y_le.setValidator(double_validator)
-        size_res_layout.addWidget(self.height_size_y_le)
-        size_res_layout.addWidget(QLabel(" Resolution:"))
-        self.height_res_x_le = QLineEdit(str(hm_default["res_x"]))
-        self.height_res_x_le.setFixedWidth(50)
-        self.height_res_x_le.setPlaceholderText("res_x")
-        int_validator = QIntValidator(1, 10000)
-        self.height_res_x_le.setValidator(int_validator)
-        size_res_layout.addWidget(self.height_res_x_le)
-        size_res_layout.addWidget(QLabel("×"))
-        self.height_res_y_le = QLineEdit(str(hm_default["res_y"]))
-        self.height_res_y_le.setFixedWidth(50)
-        self.height_res_y_le.setPlaceholderText("res_y")
-        self.height_res_y_le.setValidator(int_validator)
-        size_res_layout.addWidget(self.height_res_y_le)
-        height_layout.addRow("Size (m):", size_res_layout)
-        height_group.setLayout(height_layout)
-        self._inner_layout.addWidget(height_group)
-
-        # ---------------- Command ----------------
-        command_group = QGroupBox("Command")
-        command_layout = QFormLayout()
-
-        # command_dim (1~6)
-        self.command_dim_cb = NoWheelComboBox()
-        self.command_dim_cb.addItems([str(i) for i in range(1, 7)])
-        self.command_dim_cb.setCurrentText(str(cmd_dim_val))
-        command_layout.addRow("Command Dim:", self.command_dim_cb)
-
-        # Per-index scales grid
-        self.command_scales_group = QGroupBox("Command Scales")
-        self.command_scales_grid = QGridLayout(self.command_scales_group)
-        self.command_scales_grid.setColumnStretch(1, 1)
-        self._rebuild_command_scales(command_scales_from_cfg, int(self.command_dim_cb.currentText()))
-        self.command_dim_cb.currentTextChanged.connect(
-            lambda _:
-                (self._rebuild_command_scales(command_scales_from_cfg, int(self.command_dim_cb.currentText())),
-                 QTimer.singleShot(0, self._recalculate_height))
-        )
-
-        command_layout.addRow(self.command_scales_group)
-        command_group.setLayout(command_layout)
-        self._inner_layout.addWidget(command_group)
-
-        # Put the fully built inner widget into the scroll area
-        self._scroll.setWidget(self._inner_widget)
-
-        # ---------------- Dialog buttons (fixed at the bottom outside the scroll) ----------------
-        self._buttons = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        self._buttons.accepted.connect(self.accept)
-        self._buttons.rejected.connect(self.reject)
-        self._outer_layout.addWidget(self._buttons)
-
-        # ---------------- Populate rows after widgets are ready ----------------
-        # Priority: follow current settings' order + each obs's freq/scale
-        stacked_list_set = self.settings.get("stacked_obs_order", []) or []
-        non_stacked_list_set = self.settings.get("non_stacked_obs_order", []) or []
-
-        def scale_for_default(obs_name):
-            return to_float(obs_scales.get(obs_name, 1.0), 1.0)
-
-        if stacked_list_set or non_stacked_list_set:
-            # Restore based on current settings
-            for obs in stacked_list_set:
-                prev = self.settings.get(obs) or {}
-                self.add_stacked(
-                    obs,
-                    to_int(prev.get("freq", 50), 50),
-                    to_float(prev.get("scale", scale_for_default(obs)))
-                )
-            for obs in non_stacked_list_set:
-                prev = self.settings.get(obs) or {}
-                self.add_non(
-                    obs,
-                    to_int(prev.get("freq", 50), 50),
-                    to_float(prev.get("scale", scale_for_default(obs)))
-                )
-        else:
-            # YAML defaults
-            yaml_stacked = env_cfg.get("stacked_obs_order", []) or {}
-            yaml_non = env_cfg.get("non_stacked_obs_order", []) or {}
-            for obs in yaml_stacked:
-                self.add_stacked(obs, 50, scale_for_default(obs))
-            for obs in yaml_non:
-                self.add_non(obs, 50, scale_for_default(obs))
-
-        # Initial sizing after building widgets (final pass will run in showEvent)
-        self._recalculate_height()
-
-    def showEvent(self, event):
-        # Ensure final sizing after the dialog becomes visible
-        super().showEvent(event)
-        QTimer.singleShot(0, self._recalculate_height)
-
-    def _recalculate_height(self):
-        """Recalculate and apply dialog height to include content + buttons + margins + frame."""
-        try:
-            # Activate layouts so that size hints are accurate
-            if self._outer_layout:
-                self._outer_layout.activate()
-            if self._inner_layout:
-                self._inner_layout.activate()
-            if self.layout():
-                self.layout().activate()
-
-            # Measure inner content
-            self._inner_widget.adjustSize()
-            content_h = self._inner_widget.sizeHint().height()
-
-            # Buttons/margins/chrome
-            btn_h = self._buttons.sizeHint().height() if self._buttons else 0
-            margins = self._outer_layout.contentsMargins() if self._outer_layout else self.contentsMargins()
-            chrome_h = (margins.top() + margins.bottom() + btn_h +
-                        (self._outer_layout.spacing() if self._outer_layout else 0))
-
-            # Titlebar/frame height
-            try:
-                frame_h = self.style().pixelMetric(QStyle.PM_TitleBarHeight) \
-                          + 2 * self.style().pixelMetric(QStyle.PM_DefaultFrameWidth)
-            except Exception:
-                frame_h = 40
-
-            total_h = int(content_h + chrome_h + frame_h)
-
-            # Clamp to 90% of available screen height
-            scr = (self.screen() if hasattr(self, "screen") else None) or QApplication.primaryScreen()
-            avail_h = scr.availableGeometry().height() if scr else 900
-            max_h = int(avail_h * 0.90)
-
-            target_h = max(400, min(max_h, total_h))
-            self.resize(900, target_h)
-        except Exception:
-            # Fallback to a safe default if measurement fails
-            scr = (self.screen() if hasattr(self, "screen") else None) or QApplication.primaryScreen()
-            avail_h = scr.availableGeometry().height() if scr else 900
-            self.resize(900, int(avail_h * 0.90))
-
-    # ------- Command scale helpers -------
-    def _rebuild_command_scales(self, command_scales_from_cfg: dict, cmd_dim: int):
-        # Clear existing grid content
-        while self.command_scales_grid.count():
-            item = self.command_scales_grid.takeAt(0)
-            w = item.widget()
-            if w:
-                w.deleteLater()
-        self.cmd_scale_cbs.clear()
-
-        self.command_scales_grid.addWidget(QLabel("Index"), 0, 0)
-        self.command_scales_grid.addWidget(QLabel("Scale"), 0, 1)
-
-        # Reflect current settings with highest priority
-        prior = normalize_numkey_float_values(self.settings.get("command_scales", {}))
-
-        for i in range(cmd_dim):
-            idx_str = str(i)
-            default_val = to_float(prior.get(idx_str, command_scales_from_cfg.get(idx_str, 1.0)), 1.0)
-
-            row = i + 1
-            self.command_scales_grid.addWidget(QLabel(f"{i}"), row, 0)
-
-            cb = NoWheelComboBox()
-            cb.addItems(self._scale_options())
-            if str(default_val) in self._scale_options():
-                cb.setCurrentText(str(default_val))
-            else:
-                cb.setCurrentText("1.0")
-            self.command_scales_grid.addWidget(cb, row, 1)
-            self.cmd_scale_cbs.append(cb)
-
-    # ------- Observation rows -------
-    def add_stacked(self, selected="", freq=50, scale=1.0):
-        # Add one stacked-observation row with type/freq/scale and a delete button
-        h = QHBoxLayout()
-        combo = NoWheelComboBox()
-        combo.addItems(self.obs_types)
-        if selected:
-            combo.setCurrentText(selected)
-        h.addWidget(combo)
-
-        h.addWidget(QLabel("Freq:"))
-        freq_cb = NoWheelComboBox()
-        freq_cb.addItems(["10", "25", "50"])
-        freq_cb.setCurrentText(str(freq))
-        h.addWidget(freq_cb)
-
-        h.addWidget(QLabel("Scale:"))
-        default_scale = self.get_default_scale(selected or combo.currentText())
-        scale_cb = NoWheelComboBox()
-        scale_cb.addItems(self._scale_options())
-        items = [scale_cb.itemText(j) for j in range(scale_cb.count())]
-        if str(scale) in items:
-            scale_cb.setCurrentText(str(scale))
-        else:
-            scale_cb.setCurrentText(str(default_scale))
-        h.addWidget(scale_cb)
-
-        remove = QPushButton("Delete")
-        remove.clicked.connect(lambda: self.remove_layout(h, self.stacked_container, self.stacked_rows))
-        h.addWidget(remove)
-
-        self.stacked_container.addLayout(h)
-        self.stacked_rows.append({"layout": h, "combo": combo, "freq": freq_cb, "scale": scale_cb})
-
-        # Recalculate height after dynamic content change
-        QTimer.singleShot(0, self._recalculate_height)
-
-    def add_non(self, selected="", freq=50, scale=1.0):
-        # Add one non-stacked-observation row with type/freq/scale and a delete button
-        h = QHBoxLayout()
-        combo = NoWheelComboBox()
-        combo.addItems(self.obs_types)
-        if selected:
-            combo.setCurrentText(selected)
-        h.addWidget(combo)
-
-        h.addWidget(QLabel("Freq:"))
-        freq_cb = NoWheelComboBox()
-        freq_cb.addItems(["10", "25", "50"])
-        freq_cb.setCurrentText(str(freq))
-        h.addWidget(freq_cb)
-
-        h.addWidget(QLabel("Scale:"))
-        default_scale = self.get_default_scale(selected or combo.currentText())
-        scale_cb = NoWheelComboBox()
-        scale_cb.addItems(self._scale_options())
-        items = [scale_cb.itemText(j) for j in range(scale_cb.count())]
-        if str(scale) in items:
-            scale_cb.setCurrentText(str(scale))
-        else:
-            scale_cb.setCurrentText(str(default_scale))
-        h.addWidget(scale_cb)
-
-        remove = QPushButton("Delete")
-        remove.clicked.connect(lambda: self.remove_layout(h, self.non_container, self.non_rows))
-        h.addWidget(remove)
-
-        self.non_container.addLayout(h)
-        self.non_rows.append({"layout": h, "combo": combo, "freq": freq_cb, "scale": scale_cb})
-
-        # Recalculate height after dynamic content change
-        QTimer.singleShot(0, self._recalculate_height)
-
-    def get_default_scale(self, obs_type):
-        # Read default scale from the current environment's obs_scales
-        env_id = self.parent_widget.env_id_cb.currentText()
-        obs_scales = (self.parent_widget.env_config.get(env_id, {}) or {}).get("obs_scales", {}) or {}
-        return to_float(obs_scales.get(obs_type, 1.0), 1.0)
-
-    def remove_layout(self, layout, container, row_store):
-        # Remove the given row layout and its widgets, and update the store
-        while layout.count():
-            item = layout.takeAt(0)
-            w = item.widget()
-            if w:
-                w.deleteLater()
-        container.removeItem(layout)
-        layout.deleteLater()
-        for i, row in enumerate(row_store):
-            if row["layout"] is layout:
-                row_store.pop(i)
-                break
-
-        # Recalculate height after dynamic content change
-        QTimer.singleShot(0, self._recalculate_height)
-
-    def _extract_height_map_freq_scale_from_rows(self):
-        """From the rows, find the first 'height_map' selection and return its (freq, scale)."""
-        for rows in (self.stacked_rows, self.non_rows):
-            for row in rows:
-                if row["combo"].currentText() == "height_map":
-                    freq_val = to_int(row["freq"].currentText(), 50)
-                    scale_val = to_float(row["scale"].currentText(), 1.0)
-                    return True, freq_val, scale_val
-        return False, None, None
-
-    def get_settings(self):
-        # Aggregate dialog selections into a settings dictionary
-        stacked_order = []
-        non_order = []
-        obs_dict = {}
-
-        for row in self.stacked_rows:
-            combo = row["combo"]
-            freq_cb = row["freq"]
-            scale_cb = row["scale"]
-            obs_type = combo.currentText()
-            if obs_type:
-                stacked_order.append(obs_type)
-                obs_dict[obs_type] = {"freq": int(freq_cb.currentText()), "scale": float(scale_cb.currentText())}
-
-        for row in self.non_rows:
-            combo = row["combo"]
-            freq_cb = row["freq"]
-            scale_cb = row["scale"]
-            obs_type = combo.currentText()
-            if obs_type:
-                non_order.append(obs_type)
-                obs_dict[obs_type] = {"freq": int(freq_cb.currentText()), "scale": float(scale_cb.currentText())}
-
-        for obs_type in self.obs_types:
-            if obs_type not in obs_dict:
-                obs_dict[obs_type] = None
-
-        # Save command_scales
-        command_scales = {str(i): float(cb.currentText()) for i, cb in enumerate(self.cmd_scale_cbs)}
-
-        # Height map size/res
-        sx = to_float(self.height_size_x_le.text(), 1.0)
-        sy = to_float(self.height_size_y_le.text(), 0.6)
-        rx = to_int(self.height_res_x_le.text(), 15)
-        ry = to_int(self.height_res_y_le.text(), 9)
-
-        hm_selected, hm_freq, hm_scale = self._extract_height_map_freq_scale_from_rows()
-        if hm_selected:
-            height_map = {"size_x": sx, "size_y": sy, "res_x": rx, "res_y": ry, "freq": hm_freq, "scale": hm_scale}
-        else:
-            height_map = None
-
-        # Remove height_map from obs_dict if present (avoid overwriting the detailed dict above)
-        obs_dict.pop("height_map", None)
-
-        stack_size = int(self.stack_size_cb.currentText())
-        return {
-            "stacked_obs_order": stacked_order,
-            "non_stacked_obs_order": non_order,
-            "stack_size": stack_size,
-            "command_dim": int(self.command_dim_cb.currentText()),
-            "command_scales": command_scales,
-            "height_map": height_map,
-            **obs_dict
-        }
-
-# ---------------------------
-# Worker
-# ---------------------------
-
-class TesterWorker(QObject):
-    finished = pyqtSignal()
-    error = pyqtSignal(str)
-    def __init__(self, tester: Tester):
-        super().__init__()
-        self.tester = tester
-    def run(self):
-        # Execute tester in a separate thread context
-        try:
-            self.tester.init_user_command()
-            self.tester.test()
-        except Exception as e:
-            self.error.emit(str(e))
-        finally:
-            self.finished.emit()
-
-# ---------------------------
-# Main Window
-# ---------------------------
 
 class MainWindow(QMainWindow):
     def __init__(self):
@@ -615,8 +40,8 @@ class MainWindow(QMainWindow):
         self._last_run_had_error = False
 
     def _init_window(self):
-        app_logo_path = os.path.join(os.path.dirname(__file__), "icon", "main_logo_128_128.png")
-        self.setWindowIcon(QIcon(app_logo_path))
+        app_icon_path = os.path.join(os.path.dirname(__file__), "icon", "window_icon.png")
+        self.setWindowIcon(QIcon(app_icon_path))
         self.setWindowTitle("cosim - v1.5.0")
         self.resize(750, 970)
         self.installEventFilter(self)
